requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-mysql==3.9.0
djangorestframework==3.12.4
idna==2.10
multidict==5.1.0
mysqlclient==2.0.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
